+++ monero-rpc/src/lib.rs
@@ -14,3 +14,5 @@
 
 pub mod monerod;
 pub mod wallet;
+
+pub use jsonrpc_client as jsonrpc;